+++ src/main/java/ru/netology/model/Post.java
@@ -1,30 +1,30 @@
 package ru.netology.model;
 
 public class Post {
-    private long id;
-    private String content;
+  private long id;
+  private String content;
 
-    public Post() {
-    }
+  public Post() {
+  }
 
-    public Post(long id, String content) {
-        this.id = id;
-        this.content = content;
-    }
+  public Post(long id, String content) {
+    this.id = id;
+    this.content = content;
+  }
 
-    public long getId() {
-        return id;
-    }
+  public long getId() {
+    return id;
+  }
 
-    public void setId(long id) {
-        this.id = id;
-    }
+  public void setId(long id) {
+    this.id = id;
+  }
 
-    public String getContent() {
-        return content;
-    }
+  public String getContent() {
+    return content;
+  }
 
-    public void setContent(String content) {
-        this.content = content;
-    }
+  public void setContent(String content) {
+    this.content = content;
+  }
 }
