+++ src/main/java/ru/netology/repository/PostRepository.java
@@ -1,44 +1,15 @@
 package ru.netology.repository;
 
-import ru.netology.exception.NotFoundException;
 import ru.netology.model.Post;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 
-// Stub
-public class PostRepository {
-    private long counter;
-    private ConcurrentHashMap<Long, Post> posts = new ConcurrentHashMap<>();
+public interface PostRepository {
+    List<Post> all();
 
-    public List<Post> all() {
-        return new ArrayList<>(posts.values());
-    }
+    Optional<Post> getById(long id);
 
-    public Optional<Post> getById(long id) {
-        return Optional.ofNullable(posts.get(id));
-    }
+    Post save(Post post);
 
-    public Post save(Post post) throws NotFoundException {
-        if (post.getId() == 0) {
-            counter++;
-            post.setId(counter);
-            posts.put(counter, post);
-            return posts.get(counter);
-        } else if (posts.containsKey(post.getId())) {
-            posts.replace(post.getId(), post);
-            return posts.get(post.getId());
-        } else {
-            throw new NotFoundException("Id not found!");
-        }
-    }
-
-    public void removeById(long id) {
-        if (posts.containsKey(id)) {
-            posts.remove(id);
-        } else {
-            throw new NotFoundException("Id not found!");
-        }
-    }
+    void removeById(long id);
 }
