+++ src/main/java/ru/netology/service/PostService.java
@@ -1,31 +1,34 @@
 package ru.netology.service;
 
+import org.springframework.stereotype.Service;
 import ru.netology.exception.NotFoundException;
 import ru.netology.model.Post;
-import ru.netology.repository.PostRepository;
+import ru.netology.repository.PostRepositoryStubImpl;
 
 import java.util.List;
 
+@Service
 public class PostService {
-    private final PostRepository repository;
+  private final PostRepositoryStubImpl repository;
 
-    public PostService(PostRepository repository) {
-        this.repository = repository;
-    }
+  public PostService(PostRepositoryStubImpl repository) {
+    this.repository = repository;
+  }
 
-    public List<Post> all() {
-        return repository.all();
-    }
+  public List<Post> all() {
+    return repository.all();
+  }
 
-    public Post getById(long id) {
-        return repository.getById(id).orElseThrow(NotFoundException::new);
-    }
+  public Post getById(long id) {
+    return repository.getById(id).orElseThrow(NotFoundException::new);
+  }
 
-    public Post save(Post post) throws NotFoundException {
-        return repository.save(post);
-    }
+  public Post save(Post post) throws NotFoundException {
+    return repository.save(post);
+  }
 
-    public void removeById(long id) throws NotFoundException {
-        repository.removeById(id);
-    }
+  public void removeById(long id) throws NotFoundException {
+    repository.removeById(id);
+  }
 }
+
