+++ src/main/java/ru/netology/servlet/MainServlet.java
@@ -1,28 +1,28 @@
 package ru.netology.servlet;
 
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import ru.netology.controller.PostController;
-import ru.netology.exception.NotFoundException;
-import ru.netology.repository.PostRepository;
-import ru.netology.service.PostService;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
 
 public class MainServlet extends HttpServlet {
     private final String GET = "GET";
     private final String POST = "POST";
     private final String DELETE = "DELETE";
+    private final String PACKAGE = "ru.netology";
     private final String PATH = "/api/posts";
+    private final String CONTROLLER = "postController";
     private PostController controller;
 
     @Override
     public void init() {
-        final var repository = new PostRepository();
-        final var service = new PostService(repository);
-        controller = new PostController(service);
+        // отдаём список пакетов, в которых нужно искать аннотированные классы
+        final var context = new AnnotationConfigApplicationContext(PACKAGE);
+
+        // получаем по имени бина
+        controller = (PostController) context.getBean(CONTROLLER);
     }
 
     @Override
